% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastupgma.R
\name{fastupgma}
\alias{fastupgma}
\title{fastupgma}
\usage{
fastupgma(D, method = "average", ...)
}
\arguments{
\item{D}{A distance matrix object to create the phylogenetic tree}

\item{method}{The agglomeration method to be used. This should be
(an unambiguous abbreviation of) one of "ward", "single", "complete",
"average", "mcquitty", "median" or "centroid". The default is "average".}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A phylogenetic tree of class phylo.
}
\description{
UPGMA and WPGMA clustering.
Just a wrapper function around hclust from fastcluster package
instead base hcluse.
}
\examples{
fastupgma(CrociduraDistance)
fastupgma(rodentiaDistance, method = "mcquitty")
}
