% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrociduraDistance.R, R/rodentiaDistance.R
\docType{data}
\name{rodentiaDistance}
\alias{rodentiaDistance}
\alias{CrociduraDistance}
\title{Phylogenetic distance between Crocidura rodents}
\format{
A matrix with 192 rows and 192 cols

A matrix with 114 rows and 114 cols
}
\source{
\url{https://doi.org/10.1371/journal.pbio.3000494}

\url{https://doi.org/10.1371/journal.pbio.3000494}
}
\usage{
data(rodentiaDistance)

data(rodentiaDistance)
}
\description{
A dataset containing the phylogenetic distance matrix between Crocidura
rodents according to Upham et al 2019.

A dataset containing the phylogenetic distance matrix between rodents
according to Upham et al 2019.
}
\references{
\url{https://doi.org/10.1371/journal.pbio.3000494}

\url{https://doi.org/10.1371/journal.pbio.3000494}
}
\keyword{datasets}
